training/requirements.txt
torch
onnx
onnxruntime
python-chess
tqdm
numpy
